requirements.txt
python-binance==1.0.19
requests==2.32.3
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
ta==0.11.0
websocket-client==1.8.0